refactor: Move journal entry equality methods to base class
