+++ src/JournalCli/Core/JournalEntryFile.cs
@@ -5,7 +5,7 @@ using NodaTime;
 
 namespace JournalCli.Core
 {
-    public class JournalEntryFile : IJournalEntry, IComparable<JournalEntryFile>, IEquatable<JournalEntryFile>
+    public class JournalEntryFile : JournalEntryBase, IComparable<JournalEntryFile>
     {
         private readonly IJournalReader _reader;
 
@@ -19,40 +19,22 @@ namespace JournalCli.Core
             EntryName = reader.EntryName;
         }
 
-        public LocalDate EntryDate { get; }
+        public override LocalDate EntryDate { get; }
+
         public string FilePath { get; }
+
         public string Body { get; }
-        public string EntryName { get; }
-        public IReadOnlyCollection<string> Tags { get; }
-        public override string ToString() => EntryName;
-        public IJournalReader GetReader() => _reader;
+
+        public override string EntryName { get; }
+
+        public override IReadOnlyCollection<string> Tags { get; }
+
+        public override IJournalReader GetReader() => _reader;
 
         public int CompareTo(JournalEntryFile other)
         {
             if (ReferenceEquals(this, other)) return 0;
             return ReferenceEquals(null, other) ? 1 : EntryDate.CompareTo(other.EntryDate);
-        }
-
-        public bool Equals(JournalEntryFile other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-            return EntryDate.Equals(other.EntryDate) && string.Equals(EntryName, other.EntryName);
-        }
-
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((JournalEntryFile)obj);
-        }
-
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                return (EntryDate.GetHashCode() * 397) ^ EntryName.GetHashCode();
-            }
         }
     }
 }
